ops/requirements.txt
python-telegram-bot==22.7
python-dotenv==1.2.2
claude-agent-sdk==0.1.63
